lsgb: Print devices in proper hierarchy
The devices are currently listed in the alphabetical order in which they
are present in /sys/bus/greybus/devices/ directory, but that's not the
right way of doing it. For example, this is how it looks right now:

	$ lsgb

	  Module 08 ic:1
	    Intf 08 0xfffe0001:0xff980067 v0.0 sn:0xffffffff60400074
	      Bundle 01 Class 0x16 (Firmware Management)
	      Control (Project Ara / Second stage boot loader Interface)
	  SVC 01
	Bus 01

The Module should be listed within Bus, but it isn't. If we have
multiple modules and busses, then the busses will be listed at the
bottom and it will be very difficult to get the hierarchy right.

This patch changes this behavior and stars listing them based on the
parent directory instead. This is how it looks after this patch:

	$ lsgb

	Bus 01
	  Module 08 ic:1
	    Intf 08 0xfffe0001:0xff980067 v0.0 sn:0xffffffff60400074
	      Bundle 01 Class 0x16 (Firmware Management)
	      Control (Project Ara / Second stage boot loader Interface)
	  SVC 01

Tested on EVT1.5 and gbsim.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>